Ops/fin-cron-data/requirements_cron.txt
PyMySQL==1.0.2
python-dotenv==0.21.1
# pandas==1.5.3
pandas<2.2
SQLAlchemy==1.4.46
yfinance
# getFamaFrenchFactors
# freeoptionschain
# polygon
# polygon-api-client
# interactive-broker-python-web-api